requirements.txt
anyio==3.5.0
argon2-cffi-bindings==21.2.0
argon2-cffi==21.3.0
asttokens==2.0.5
attrs==21.4.0
babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.11.1
biopython==1.78
biotite==0.34.1
bleach==4.1.0
bottleneck==1.3.5
brotlipy==0.7.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.3
cryptography==37.0.1
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
entrypoints==0.4
executing==0.8.3
fair-esm==1.0.3
fastjsonschema==2.16.2
fonttools==4.25.0
gitdb==4.0.9
gitpython==3.1.29
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.2.0
ipykernel==6.15.2
ipython-genutils==0.2.0
ipython==8.4.0
jedi==0.18.1
jinja2==3.0.3
joblib==1.1.0
json5==0.9.6
jsonschema==4.16.0
jupyter-client==7.3.5
jupyter-core==4.11.1
jupyter-server==1.18.1
jupyterlab-pygments==0.1.2
jupyterlab-server==2.15.2
jupyterlab==3.4.4
kiwisolver==1.4.2
markupsafe==2.1.1
matplotlib-inline==0.1.6
matplotlib==3.5.2
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
msgpack==1.0.4
munkres==1.1.4
nbclassic==0.3.5
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.5.0
nest-asyncio==1.5.5
networkx==2.8.7
notebook==6.4.12
numexpr==2.8.3
numpy==1.23.1
packaging==21.3
pandas==1.4.4
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==9.2.0
pip==22.2.2
pkgutil-resolve-name==1.3.10
ply==3.11
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.20
protobuf==4.21.7
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pygments==2.11.2
pyopenssl==22.0.0
pyparsing==3.0.9
pyqt5-sip==12.11.0
pyrsistent==0.18.0
pysocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
pyyaml==6.0
pyzmq==23.2.0
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.11.2
send2trash==1.8.0
sentry-sdk==1.9.10
setproctitle==1.3.2
setuptools==63.4.1
shortuuid==1.0.9
sip==6.6.2
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.1
stack-data==0.2.0
terminado==0.13.1
testpath==0.6.0
threadpoolctl==2.2.0
toml==0.10.2
torch-cluster==1.6.0
torch-geometric==2.1.0.post1
torch-scatter==2.0.9
torch-sparse==0.6.15
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.1.1
typing-extensions==4.3.0
urllib3==1.26.11
wandb==0.13.4
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
wheel==0.37.1
zipp==3.8.0